# www/api-transfer/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger-Demo",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
        "/purses": {
            "post": {
                "tags": [
                    "Purses"
                ],
                "summary": "Create purse",
                "description": "Return created new purse",
                "operationId": "store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "balance": {
                                        "type": "number"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "data of new purse",
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "number"
                                                },
                                                "balance": {
                                                    "type": "number"
                                                },
                                                "status": {
                                                    "type": "string"
                                                },
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "number"
                                                        },
                                                        "name": {
                                                            "type": "string"
                                                        },
                                                        "email": {
                                                            "type": "string"
                                                        },
                                                        "document": {
                                                            "type": "string"
                                                        },
                                                        "profile": {
                                                            "type": "string"
                                                        },
                                                        "status": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "The given data was invalid",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Server Erro",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/purses/{purse_id}/balance": {
            "patch": {
                "tags": [
                    "Purses"
                ],
                "summary": "Update Balance from Purse",
                "description": "Return updated Balance",
                "operationId": "updateBalance",
                "parameters": [
                    {
                        "name": "purse_id",
                        "in": "path",
                        "description": "Purse id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "balance": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "data of updated wallet",
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "number"
                                                },
                                                "balance": {
                                                    "type": "number"
                                                },
                                                "status": {
                                                    "type": "string"
                                                },
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "number"
                                                        },
                                                        "name": {
                                                            "type": "string"
                                                        },
                                                        "email": {
                                                            "type": "string"
                                                        },
                                                        "document": {
                                                            "type": "string"
                                                        },
                                                        "profile": {
                                                            "type": "string"
                                                        },
                                                        "status": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "The given data was invalid",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Server Erro",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}